Implement the MediaPlayer playback in the service instead of fragment.

+++ app/src/main/java/dengn/spotifystreamer/utils/PlayerUtils.java
@@ -77,7 +77,7 @@ public class PlayerUtils {
      * @param totalDuration
      * returns current duration in milliseconds
      * */
-    public static int progressToTimer(int progress, int totalDuration) {
+    public static int progressToTimer(int progress, long totalDuration) {
         int currentDuration = 0;
         totalDuration = (int) (totalDuration / 1000);
         currentDuration = (int) ((((double)progress) / 100) * totalDuration);
